% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{antifreeze}
\alias{antifreeze}
\title{Water content of antifreeze data}
\format{
A data frame with 34 rows and 1 column:
\describe{
\item{water}{water content, in ppm}
}
}
\source{
Wetherill, G.B. and Brown, D.W. (1991) Statistical Process
Control. New York: Chapman & Hall, p. 120
}
\usage{
antifreeze
}
\description{
Water content in 34 successive batches of antifreeze.
}
\seealso{
Other datasets in \code{qcc} package:
\code{\link{RyanMultivar}},
\code{\link{boiler}},
\code{\link{circuit}},
\code{\link{dyedcloth}},
\code{\link{orangejuice}},
\code{\link{pcmanufact}},
\code{\link{pistonrings}},
\code{\link{viscosity}}
}
\concept{datasets in \code{qcc} package}
\keyword{datasets}
