% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{blues.colors}
\alias{blues.colors}
\title{Blue color palette}
\usage{
blues.colors(n)
}
\arguments{
\item{n}{the number of colors.}
}
\value{
A character vector containing the requested colors.
}
\description{
Creates a palette of blue colors.
}
