% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boiler}
\alias{boiler}
\title{Boiler temperature data}
\format{
A data frame with 25 observations on the following 8 variables:
\describe{
\item{t1}{temperature reading 1}
\item{t2}{temperature reading 2}
\item{t3}{temperature reading 3}
\item{t4}{temperature reading 4}
\item{t5}{temperature reading 5}
\item{t6}{temperature reading 6}
\item{t7}{temperature reading 7}
\item{t8}{temperature reading 8}
}
}
\source{
Mason, R.L. and Young, J.C. (2002) \emph{Multivariate
Statistical Process Control with Industrial Applications}, SIAM, p. 86.
}
\usage{
boiler
}
\description{
Temperature readings from the eight configured burners on a boiler.
}
\seealso{
Other datasets in \code{qcc} package:
\code{\link{RyanMultivar}},
\code{\link{antifreeze}},
\code{\link{circuit}},
\code{\link{dyedcloth}},
\code{\link{orangejuice}},
\code{\link{pcmanufact}},
\code{\link{pistonrings}},
\code{\link{viscosity}}
}
\concept{datasets in \code{qcc} package}
\keyword{datasets}
