% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{c4}
\alias{c4}
\title{The \eqn{c_4}{\code{c4}} Constant}
\usage{
c4(n)
}
\arguments{
\item{n}{A vector of sample sizes.}
}
\value{
A vector of calculated \eqn{c_4}{c4} constants.
}
\description{
Calculates the bias-correction factor for the sample standard deviation.
}
\details{
For a normally distributed sample of size \eqn{n}, the sample standard
deviation \eqn{S} is biased downward,
with \deqn{E(S) = c_4(n)\sigma.}{E(s) = c4(n) * sigma}
Consequently, \eqn{S / c_4(n)}{S/c4(n)} is an unbiased estimator of the population
standard deviation.
}
\seealso{
For other implementations in R:
\code{\link[SixSigma:ss.cc.getc4]{SixSigma::ss.cc.getc4()}}, \code{\link[IQCC:c4]{IQCC::c4()}}, \code{\link[shewhartr:shewhart_constants]{shewhartr::shewhart_constants()}}

Other constants for Shewhart charts:
\code{\link[=d2]{d2()}},
\code{\link[=d3]{d3()}},
\code{\link[=d4]{d4()}}
}
\concept{constants for Shewhart charts}
