% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causeeffect.R
\name{causeEffectDiagram}
\alias{causeEffectDiagram}
\title{Cause and Effect Diagram}
\usage{
causeEffectDiagram(
  cause,
  effect,
  title = "Cause-and-Effect diagram",
  cex = c(0.9, 1, 1.2),
  font = c(3, 1, 2),
  ...
)
}
\arguments{
\item{cause}{a list of causes and branches providing descriptive labels (see
the example below).}

\item{effect}{a string label or the effect.}

\item{title}{a character string specifying the main title. Set \code{title = NULL} to remove the title.}

\item{cex}{a vector of values for the graphical character expansion. The
values refer, in order, to branches, causes and effect.}

\item{font}{a vector of values for the font to use. The values refer, in
order, to branches, causes and effect.}

\item{...}{catches further ignored arguments.}
}
\description{
Draw a basic cause and effect diagram.
}
\examples{

causeEffectDiagram(cause = list(Measurements = c("Micrometers", 
                                                 "Microscopes", 
                                                 "Inspectors"),
                                Materials = c("Alloys", 
                                              "Lubricants", 
                                              "Suppliers"),
                                Personnel = c("Shifts", 
                                              "Supervisors", 
                                              "Training", 
                                              "Operators"),
                                Environment = c("Condensation", 
                                                "Moisture"),
                                Methods = c("Brake",
                                            "Engager", 
                                            "Angle"),
                                Machines = c("Speed", 
                                             "Lathes", 
                                             "Bits", 
                                             "Sockets")),
                   effect = "Surface Flaws")

}
\references{
\itemize{
\item Montgomery, D.C. (2013). \emph{Introduction to Statistical Quality Control}, 7th ed. New York: John Wiley & Sons.
\item Wetherill, G. B., and Brown, D. W. (1991). \emph{Statistical Process Control}. New York: Chapman & Hall.
}
}
\author{
Luca Scrucca
}
