% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-layer.R
\name{chart_common}
\alias{chart_common}
\title{Shared Control Chart (Shewhart, EWMA, CUSUM) Documentation}
\arguments{
\item{data}{a data frame, a matrix or a vector containing observed data for
the variable to chart. Each row of a data frame or a matrix, and each value
of a vector, refers to a sample or ''rationale group''.}

\item{newdata}{a data frame, matrix or vector, as for the \code{data}
argument, providing further data to plot but not included in the
computations.}

\item{newsizes}{a vector as for the \code{sizes} argument providing further
data sizes to plot but not included in the computations.}

\item{center}{a value specifying the center of group statistics or target.}
}
\description{
Shared Control Chart (Shewhart, EWMA, CUSUM) Documentation
}
\keyword{internal}
