% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{circuit}
\alias{circuit}
\title{Circuit boards data}
\format{
A data frame with 46 observations on the following 4 variables:
\describe{
\item{sample}{sample number}
\item{x}{number of defectives in 100 printed circuit boards (inspection unit)}
\item{size}{sample size}
\item{trial}{trial sample indicator (TRUE/FALSE)}
}
}
\source{
Montgomery, D.C. (1991) \emph{Introduction to Statistical
Quality Control}, 2nd ed, New York, John Wiley & Sons, pp. 173--175
}
\usage{
circuit
}
\description{
Number of nonconformities observed in 26 successive samples of 100 printed
circuit boards. Sample 6 and 20 are outside the control limits. Sample 6 was
examined by a new inspector and he did not recognize several type of
nonconformities that could have been present. Furthermore, the unusually
large number of nonconformities in sample 20 resulted from a temperature
control problem in the wave soldering machine, which was subsequently
repaired. The last 20 samples are further samples collected on inspection
units (each formed by 100 boards).
}
\seealso{
Other datasets in \code{qcc} package:
\code{\link{RyanMultivar}},
\code{\link{antifreeze}},
\code{\link{boiler}},
\code{\link{dyedcloth}},
\code{\link{orangejuice}},
\code{\link{pcmanufact}},
\code{\link{pistonrings}},
\code{\link{viscosity}}
}
\concept{datasets in \code{qcc} package}
\keyword{datasets}
