% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{d2}
\alias{d2}
\title{The \eqn{d_2}{d2} Constant}
\usage{
d2(n)
}
\arguments{
\item{n}{A vector of sample sizes.}
}
\value{
A vector of calculated \eqn{d_2}{d2} constants.
}
\description{
Calculate the \eqn{d_2}{d2} constants for a vector \code{n} of sample sizes.
}
\details{
\eqn{d_2(n)}{d2(n)} is the expected value of the range of n observations
from a standard normal distribution.
\deqn{d_2(n) = \frac{E(r)}{\sigma}}{d2(n) = E(r)/sigma}
}
\references{
\itemize{
\item Cano, Emilio L., Moguerza, Javier M. and Redchuk, Andres. 2012. \emph{Six Sigma with {R}. Statistical Engineering for ProcessImprovement}, Use R!, vol. 36. Springer, New York.
}
}
\seealso{
For other implementations in R:
\code{\link[SixSigma:ss.cc.getd2]{SixSigma::ss.cc.getd2()}}, \code{\link[IQCC:d2]{IQCC::d2()}}, \code{\link[shewhartr:shewhart_constants]{shewhartr::shewhart_constants()}}

Other constants for Shewhart charts:
\code{\link[=c4]{c4()}},
\code{\link[=d3]{d3()}},
\code{\link[=d4]{d4()}}
}
\concept{constants for Shewhart charts}
