% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{d4}
\alias{d4}
\title{The \eqn{d_4}{d4} Constant}
\usage{
d4(n)
}
\arguments{
\item{n}{A vector of sample sizes.}
}
\value{
A vector of calculated \eqn{d_4}{d4} constants.
}
\description{
Calculate the \eqn{d_4}{d4} constants for a vector \code{n} of sample sizes.
}
\details{
\eqn{d_4(n)}{d4(n)} is the median of the range of \code{n} independent
observations from a standard normal distribution. For normal observations
with standard deviation \eqn{\sigma},
\deqn{\mathrm{median}(R) = d_4(n)\sigma.}{median(R) = d4(n) * sigma}

Not to be confused with the upper range-chart limit factor
\eqn{D_4 = 1 + 3d_3/d_2}{D4 = 1 + 3 * d3 / d2}.
}
\references{
\href{https://support.minitab.com/en-us/minitab/help-and-how-to/quality-and-process-improvement/control-charts/how-to/variables-charts-for-individuals/i-mr-chart/methods-and-formulas/methods-and-formulas-for-moving-range-chart/}{Minitab: Methods and formulas for Moving Range Chart}
}
\seealso{
\code{\link[stats:qtukey]{stats::qtukey()}}, \code{\link[stats:ptukey]{stats::ptukey()}}, \code{\link[=sd.xbar.one]{sd.xbar.one()}}

Other constants for Shewhart charts:
\code{\link[=c4]{c4()}},
\code{\link[=d2]{d2()}},
\code{\link[=d3]{d3()}}
}
\concept{constants for Shewhart charts}
