% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\alias{print.describe}
\title{Descriptive statistics}
\usage{
describe(data, by, detailed = FALSE, ...)

\method{print}{describe}(x, digits = getOption("digits") - 3, ...)
}
\arguments{
\item{data}{a matrix or data frame.}

\item{by}{a factor or character vector specifying the conditioning variable.
This may be a variable included in \code{data} or an object defined in the
current environment.}

\item{detailed}{a logical specifying if detailed statistics should be
provided.}

\item{...}{additional arguments to be passed to the generic function.}

\item{x}{an object of class \code{'describe'}.}

\item{digits}{the number of significant digits to use.}
}
\description{
Compute and print descriptive statistics for a matrix or data frame,
eventually conditioning on another variable.
}
\details{
This function provides simple descriptive statistics for different types of
variables, such as numeric, factor, and logical, conditioning or not on
another variable.
}
\examples{
data(warpbreaks)
describe(warpbreaks)
describe(warpbreaks, detail = TRUE)
describe(warpbreaks, by = wool)
}
\seealso{
\code{\link[=summary]{summary()}}, \code{\link[=by]{by()}}
}
\author{
Luca Scrucca
}
