% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dyedcloth}
\alias{dyedcloth}
\title{Dyed cloth data}
\format{
A data frame with 10 observations on the following 2 variables:
\describe{
\item{x}{number of nonconformities per 50 square meters (inspection units)}
\item{size}{number of inspection units in roll (variable sample size)}
}
}
\source{
Montgomery, D.C. (1991) \emph{Introduction to Statistical
Quality Control}, 2nd ed, New York, John Wiley & Sons, pp. 183--184
}
\usage{
dyedcloth
}
\description{
In a textile finishing plant, dyed cloth is inspected for the occurrence of
defects per 50 square meters. The data on ten rolls of cloth are presented.
}
\seealso{
Other datasets in \code{qcc} package:
\code{\link{RyanMultivar}},
\code{\link{antifreeze}},
\code{\link{boiler}},
\code{\link{circuit}},
\code{\link{orangejuice}},
\code{\link{pcmanufact}},
\code{\link{pistonrings}},
\code{\link{viscosity}}
}
\concept{datasets in \code{qcc} package}
\keyword{datasets}
