% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewma.R
\name{ewmaSmooth}
\alias{ewmaSmooth}
\title{EWMA smoothing function}
\usage{
ewmaSmooth(x, y, lambda = 0.2, start, ...)
}
\arguments{
\item{x}{a vector of x-values.}

\item{y}{a vector of y-values.}

\item{lambda}{the smoothing parameter.}

\item{start}{the starting value.}

\item{...}{additional arguments (currently not used).}
}
\value{
Returns a list with elements:
\itemize{
\item \code{x}: ordered x-values.
\item \code{y}: smoothed y-values.
\item \code{lambda}: the smoothing parameter.
\item \code{start}: the starting value.
}
}
\description{
Compute Exponential Weighted Moving Average.
}
\details{
EWMA function smooths a series of data based on a moving average with
weights which decay exponentially.

For each \eqn{y_t}{y_t} value the smoothed value is computed as
\deqn{z_t = \lambda y_t + (1-\lambda) z_{t-1}}
where \eqn{0 \le \lambda \le 1}{0 <= lambda <= 1} controls the weights
applied.
}
\examples{
x  = 1:50
y  = rnorm(50, sin(x/5), 0.5)
plot(x,y)
lines(ewmaSmooth(x,y,lambda=0.1), col="red")
}
\references{
\itemize{
\item Montgomery, D.C. (2013). \emph{Introduction to Statistical Quality Control}, 7th ed. New York: John Wiley & Sons.
\item Wetherill, G. B., and Brown, D. W. (1991). \emph{Statistical Process Control}. New York: Chapman & Hall.
}
}
\seealso{
\code{\link[=qcc]{qcc()}}, \code{\link[=cusum]{cusum()}}
}
\author{
Luca Scrucca
}
