% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFundamentalData.R
\name{getFundamentalData}
\alias{getFundamentalData}
\title{Retreive fundamental data from database}
\usage{
qcc <- QCC(url, jwt)
qcc$getFundamentalData(id, from, to, sec, min, hour, day, mon, dow)
}
\arguments{
\item{id}{A string representing the the product id}

\item{from}{A string in the form of "yyyy-mm-dd" representing the start date desired}

\item{to}{A string in the form of "yyyy-mm-dd" representing the end date desired}

\item{sec}{A string ranges from 00 to 59, representing the corrsponding second, if several seconds are desired, separated by ","}

\item{min}{A string ranges from 00 to 59, representing the corrsponding minute, if several minutes are desired, separated by ","}

\item{hour}{A string ranges from 00 to 23, representing the corrsponding hour, if several hours are desired, separated by ","}

\item{day}{A string ranges from 01 to 31, representing the corresponding day, if several days are desired, separated by ","}

\item{mon}{A string ranges from 01 to 12, representing the corresponding month, if several months are desired, separated by ","}

\item{dow}{A string ranges from 0 to 6, representing the corresponding day of week, is several day of week are desired, separated by ","}
}
\value{
A data frame containing the user desired data frame from the corresponding api address filterd by all given parameters of the function
}
\description{
Extract fundamental data from the QCC database corresponding to user's input, such as time period, type of time
}

