% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuoteData.R
\name{getQuotData}
\alias{getQuotData}
\title{Retrieve quote data from database}
\usage{
qcc <- QCC(url, jwt)
qcc$getQuoteData(id, freq, fields, from, to)
}
\arguments{
\item{id}{A string representing the the InstrumentID for the product}

\item{freq}{A string representing the user desired frequency of the data. It can be chosen from "1min", "5min", "15min", "Day", "Hour"}

\item{fields}{A string representing the user desired information fields of the product. If more than 1 field is desired, separate them by blank space}

\item{from}{A string in the form of "yyyy.mm.dd" representing the start date desired}

\item{to}{A string in the form of "yyyy.mm.dd" representing the end date desired}
}
\value{
A data frame containing the user desired data frame from the corresponding api address filterd by all given parameters of the function
}
\description{
It takes in paramters regrading the InstrumentID, data frequency, desired fields for data information and time period, and returns a data frame containing all desired quote data.
}

