% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orangejuice}
\alias{orangejuice}
\alias{orangejuice2}
\title{Orange juice data}
\format{
\code{orangejuice} is a data frame with 54 observations and \code{orangejuice2}
is a data frame with 64 observations. Both contain the following 4 variables:
\describe{
\item{sample}{sample id}
\item{D}{number of defectives}
\item{size}{sample sizes}
\item{trial}{trial samples (TRUE/FALSE)}
}
}
\source{
Montgomery, D.C. (1991) \emph{Introduction to Statistical
Quality Control}, 2nd ed, New York, John Wiley & Sons, pp. 152--159.
}
\usage{
orangejuice

orangejuice2
}
\description{
Frozen orange juice concentrate is packaged in 6-ounce cardboard cans. A
machine forms each can by spinning it from cardboard stock and attaching a
metal bottom panel. A can is classified as nonconforming if an inspection
indicates that it could leak when filled, either along the side seam or
around the bottom joint.
}
\details{
Thirty samples of 50 cans each were collected at half-hour intervals while
the machine operated continuously over three shifts. A new batch of cardboard
stock was introduced beginning with sample 15, and an inexperienced operator
was temporarily assigned to the machine for sample 23. After these 30
samples, the machine was adjusted and 24 additional samples were collected.
The \code{orangejuice2} dataset contains samples collected after this adjustment.
}
\seealso{
Other datasets in \code{qcc} package:
\code{\link{RyanMultivar}},
\code{\link{antifreeze}},
\code{\link{boiler}},
\code{\link{circuit}},
\code{\link{dyedcloth}},
\code{\link{pcmanufact}},
\code{\link{pistonrings}},
\code{\link{viscosity}}
}
\concept{datasets in \code{qcc} package}
\keyword{datasets}
