% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pistonrings}
\alias{pistonrings}
\title{Piston rings data}
\format{
A data frame with 200 observations on the following 3 variables:
\describe{
\item{diameter}{a numeric vector}
\item{sample}{sample ID}
\item{trial}{preliminary sample indicator (TRUE/FALSE)}
}
}
\source{
Montgomery, D.C. (1991) \emph{Introduction to Statistical
Quality Control}, 2nd ed, New York, John Wiley & Sons, pp. 206--213
}
\usage{
pistonrings
}
\description{
Piston rings for an automotive engine are produced by a forging process. The
inside diameter of the rings manufactured by the process is measured on 25
samples, each of size 5, for the control phase I, when preliminary samples
from a process being considered 'in control' are used to construct control
charts. Then, further 15 samples, again each of size 5, are obtained for
phase II.
}
\seealso{
Other datasets in \code{qcc} package:
\code{\link{RyanMultivar}},
\code{\link{antifreeze}},
\code{\link{boiler}},
\code{\link{circuit}},
\code{\link{dyedcloth}},
\code{\link{orangejuice}},
\code{\link{pcmanufact}},
\code{\link{viscosity}}
}
\concept{datasets in \code{qcc} package}
\keyword{datasets}
