% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{qcc-deprecated}
\alias{qcc-deprecated}
\alias{pareto.chart}
\alias{process.capability}
\alias{oc.curves}
\alias{qcc.overdispersion.test}
\alias{cause.and.effect}
\alias{qcc.groups}
\title{Deprecated Functions in qcc package}
\usage{
pareto.chart(...)

process.capability(...)

oc.curves(...)

qcc.overdispersion.test(...)

cause.and.effect(...)

qcc.groups(...)
}
\arguments{
\item{...}{all arguments are passed down.}
}
\description{
These functions are provided for compatibility with older versions of the
package qcc, and they will likely be removed in the future.
}
