\name{qcc-internal}
\alias{.printShortMatrix}
\alias{.qcc.options}
\alias{qcc.c4}
\alias{qccStartupMessage}

\title{Internal 'qcc' functions}

\description{Internal functions for package 'qcc'.}

\usage{
.printShortMatrix(x, head = 2, tail = 1, chead = 5, ctail = 1, \dots)

qcc.c4(n)

qccStartupMessage()
}

% \arguments{}

\details{These functions are not intended to be called by the user.}

\author{Luca Scrucca}

\keyword{internal}

\keyword{package}
