% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcc-package.R
\docType{package}
\name{qcc-package}
\alias{qcc-package}
\title{Quality Control Charts}
\description{
Shewhart quality control charts for continuous, attribute and count data.
Cusum and EWMA charts. Operating characteristic curves. Process capability
analysis. Pareto chart and cause-and-effect chart. Multivariate control
charts.
}
\details{
See \href{../doc/index.html}{vignette and documentation} accompanying the
package.
}
\references{
\itemize{
\item Scrucca, L. (2004). qcc: an R package for quality control charting and statistical process control. \emph{R News}, \strong{4}(1), 11--17.
}
}
\seealso{
\code{\link[=qcc]{qcc()}}, \code{\link[=mqcc]{mqcc()}}, \code{\link[=cusum]{cusum()}},
\code{\link[=ewma]{ewma()}}, \code{\link[=ocCurves]{ocCurves()}}, \code{\link[=processCapability]{processCapability()}},
\code{\link[=paretoChart]{paretoChart()}}, \code{\link[=causeEffectDiagram]{causeEffectDiagram()}}.
}
\author{
Luca Scrucca
}
\keyword{internal}
\keyword{package}
