% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{rules}
\alias{rules}
\alias{qccRules}
\alias{qccRulesViolatingWER1}
\alias{qccRulesViolatingWER2}
\alias{qccRulesViolatingWER3}
\alias{qccRulesViolatingWER4}
\alias{qccRulesViolatingNEL1}
\alias{qccRulesViolatingNEL2}
\alias{qccRulesViolatingNEL3}
\alias{qccRulesViolatingNEL4}
\alias{qccRulesViolatingNEL5}
\alias{qccRulesViolatingNEL6}
\alias{qccRulesViolatingNEL7}
\alias{qccRulesViolatingNEL8}
\title{Rules for Shewhart charts}
\usage{
qccRules(object, rules = object$rules, rule.set = object$rule.set)

qccRulesViolatingWER1(object, limits = object$limits)

qccRulesViolatingWER2(
  object,
  run.points = 2,
  run.length = 3,
  k = object$nsigmas * 2/3
)

qccRulesViolatingWER3(object, ...)

qccRulesViolatingWER4(object)

qccRulesViolatingNEL1(object)

qccRulesViolatingNEL2(object, run.length = 9)

qccRulesViolatingNEL3(object)

qccRulesViolatingNEL4(object)

qccRulesViolatingNEL5(object)

qccRulesViolatingNEL6(object)

qccRulesViolatingNEL7(object)

qccRulesViolatingNEL8(object)
}
\arguments{
\item{object}{an object of class \code{'qcc'}.}

\item{rules}{a vector of values specifying the rules used to declare a
process OOC or OOT. Numeric values are interpreted differently depending
on the value of \code{rule.set}.}

\item{rule.set}{a character string selecting the rule set used to interpret
numeric \code{rules} values. Use \code{"western-electric"} for Western
Electric rules or \code{"nelson"} for Nelson rules. default: \code{"western-electric"}}

\item{limits}{a two-column matrix containing the lower and upper control
limits.}

\item{run.points}{the minimum number of points beyond the sigma limit
needed to signal a violation.}

\item{run.length}{the number of consecutive points in the run.}

\item{k}{the number of standard deviations from the center line used to
define the sigma limits.}

\item{...}{additional arguments, currently ignored.}
}
\value{
The function \code{qccRules} returns a vector of the same length as the statistics
charted with value \code{NA} if no rule is violated, and numerical values
corresponding to the violated rule. If more than one rule is violated, the
value returned corresponds to the most serious violation.
}
\description{
Functions which implement rules to signal out-of-control (OOC) or
out-of-trend (OOT) points in Shewhart control charts.
}
\details{
The \code{qccRulesViolatingWER1}, \code{qccRulesViolatingWER2}, ...,
\code{qccRulesViolatingWER4} functions return the indices violating the
corresponding Western Electric rule:
\enumerate{
\item One point plots outside 3-sigma control limits.
\item Two of three consecutive points plot beyond a 2-sigma limit.
\item Four of five consecutive points plot beyond a 1-sigma limit.
\item Eight consecutive points plot on one side of the center line.
}

The \code{qccRulesViolatingNEL1}, \code{qccRulesViolatingNEL2}, ...,
\code{qccRulesViolatingNEL8} functions return the indices violating the
corresponding Nelson rule:
\enumerate{
\item One point plots outside 3-sigma control limits.
\item Nine points in a row plot on the same side of the center line.
\item Six points in a row are steadily increasing or decreasing.
\item Fourteen points in a row alternate up and down.
\item Two of three consecutive points plot beyond a 2-sigma limit on the same
side.
\item Four of five consecutive points plot beyond a 1-sigma limit on the same
side.
\item Fifteen points in a row plot within 1 sigma of the center line.
\item Eight points in a row plot outside 1 sigma on both sides of the center
line.
}
}
\author{
Luca Scrucca

Anas Sheashaey
}
