% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RyanMultivar}
\alias{RyanMultivar}
\title{Ryan's (2011) multivariate data for quality control}
\format{
Multivariate data on 20 samples of size 4 for two variables:
\describe{
\item{RyanMultivar}{a list of two data frames, \code{X1} and
\code{X2}, one for each variable.}
}
}
\source{
Ryan, T. P. (2011), \emph{Statistical Methods for Quality
Improvement}, 3rd ed. New York: John Wiley & Sons, Inc.
}
\usage{
RyanMultivar
}
\description{
Multivariate data from Ryan (2011, Table 9.2).
}
\seealso{
Other datasets in \code{qcc} package:
\code{\link{antifreeze}},
\code{\link{boiler}},
\code{\link{circuit}},
\code{\link{dyedcloth}},
\code{\link{orangejuice}},
\code{\link{pcmanufact}},
\code{\link{pistonrings}},
\code{\link{viscosity}}
}
\concept{datasets in \code{qcc} package}
\keyword{datasets}
