% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-layer.R
\name{spc_common}
\alias{spc_common}
\title{Shared Chart Methods Documentation}
\arguments{
\item{data}{The observed data values.}

\item{center}{The sample/group center statistic.}

\item{sizes}{The sample sizes.}

\item{std.dev}{The within-group standard deviation.}

\item{nsigmas}{Number of sigmas used to compute control limits.
Ignored when the \code{conf} argument is provided.}

\item{conf}{Confidence level used to compute control limits.
Must be a numeric value in \eqn{(0,1)}.}
}
\description{
Shared Chart Methods Documentation
}
\seealso{
\code{\link[=qcc]{qcc()}}
}
\author{
Luca Scrucca
}
\keyword{internal}
