% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.g.R
\name{stats.g}
\alias{stats.g}
\alias{sd.g}
\alias{limits.g}
\title{Statistics used in computing and drawing a Shewhart g chart}
\usage{
stats.g(data, sizes)

sd.g(data, sizes, ...)

limits.g(center, std.dev, sizes, nsigmas = NULL, conf = NULL)
}
\arguments{
\item{data}{The observed data values.}

\item{sizes}{sample sizes (not used)}

\item{...}{catches further ignored arguments.}

\item{center}{The sample/group center statistic.}

\item{std.dev}{standard deviation of geometric distribution}

\item{nsigmas}{Number of sigmas used to compute control limits.
Ignored when the \code{conf} argument is provided.}

\item{conf}{Confidence level used to compute control limits.
Must be a numeric value in \eqn{(0,1)}.}
}
\value{
The function \code{stats.g()} returns a list with components
\code{statistics} and \code{center}.

The function \code{sd.g()} returns \code{std.dev} the standard deviation
\eqn{sqrt(1-p)/p}.

The function \code{limits.g()} returns a matrix with lower and upper control
limits.
}
\description{
These functions are used to compute statistics required by the g chart
(geometric distribution) for use with the qcc package.
}
\details{
The g chart plots the number of non-events between events. np charts do not
work well when the probability of an event is rare (see example below).
Instead of plotting the number of events, the g chart plots the number of
non-events between events.

The geometric distribution is quite skewed so it is best to set \code{conf}
at the required confidence interval (0 < conf < 1) rather than as a
multiplier of sigma.
}
\examples{

success  = rbinom(1000, 1, 0.01)
num.noevent  = diff(which(c(1,success)==1))-1
qcc(success, type = "np", sizes = 1)
qcc(num.noevent, type = "g")
}
\references{
\itemize{
\item Kaminsky, F. C., et al. (1992). \emph{Statistical Control Charts Based on a Geometric Distribution}. \emph{Journal of Quality Technology}, \strong{24}, 63--69.
\item Yang, Z., et al. (2002). On the Performance of Geometric Charts with Estimated Control Limits. \emph{Journal of Quality Technology}, \strong{34}, 448--458.
}
}
\seealso{
\code{\link[=qcc]{qcc()}}
}
\author{
Greg Snow (greg.snow@ihc.com)
}
