% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.p.R
\name{stats.p}
\alias{stats.p}
\alias{sd.p}
\alias{limits.p}
\title{Statistics used in computing and drawing a Shewhart p chart}
\usage{
stats.p(data, sizes)

sd.p(data, sizes, ...)

limits.p(center, std.dev, sizes, nsigmas = NULL, conf = NULL)
}
\arguments{
\item{data}{The observed data values.}

\item{sizes}{The sample sizes.}

\item{...}{catches further ignored arguments.}

\item{center}{The sample/group center statistic.}

\item{std.dev}{The within-group standard deviation.}

\item{nsigmas}{Number of sigmas used to compute control limits.
Ignored when the \code{conf} argument is provided.}

\item{conf}{Confidence level used to compute control limits.
Must be a numeric value in \eqn{(0,1)}.}
}
\value{
The function \code{stats.p} returns a list with components
\code{statistics} and \code{center}.

The function \code{sd.p} returns \code{std.dev} the standard deviation of
the statistic charted.

The function \code{limits.p} returns a matrix with lower and upper control
limits.
}
\description{
These functions are used to compute statistics required by the p chart.
}
\references{
\itemize{
\item Montgomery, D.C. (2013). \emph{Introduction to Statistical Quality Control}, 7th ed. New York: John Wiley & Sons.
\item Wetherill, G. B., and Brown, D. W. (1991). \emph{Statistical Process Control}. New York: Chapman & Hall.
}
}
\seealso{
\code{\link[=qcc]{qcc()}}
}
\author{
Luca Scrucca
}
