% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.s.R
\name{stats.S}
\alias{stats.S}
\alias{sd.S}
\alias{limits.S}
\title{Functions to plot Shewhart S chart}
\usage{
stats.S(data, sizes)

sd.S(data, sizes, std.dev = c("UWAVE-SD", "MVLUE-SD", "RMSDF"), ...)

limits.S(center, std.dev, sizes, nsigmas = NULL, conf = NULL)
}
\arguments{
\item{data}{The observed data values.}

\item{sizes}{samples sizes. Optional}

\item{std.dev}{within group standard deviation. Optional for \code{sd.S}
function, required for \code{limits.S}. See \code{\link[=sd.xbar]{sd.xbar()}}.}

\item{...}{catches further ignored arguments.}

\item{center}{The sample/group center statistic.}

\item{nsigmas}{Number of sigmas used to compute control limits.
Ignored when the \code{conf} argument is provided.}

\item{conf}{Confidence level used to compute control limits.
Must be a numeric value in \eqn{(0,1)}.}
}
\value{
The function \code{stats.S} returns a list with components
\code{statistics} and \code{center}.

The function \code{sd.S} returns \code{std.dev} the standard deviation of
the statistic charted.

The function \code{limits.S} returns a matrix with lower and upper control
limits.
}
\description{
These functions are used to compute statistics required by the S chart.
}
\references{
\itemize{
\item Montgomery, D.C. (2013). \emph{Introduction to Statistical Quality Control}, 7th ed. New York: John Wiley & Sons.
\item Wetherill, G. B., and Brown, D. W. (1991). \emph{Statistical Process Control}. New York: Chapman & Hall.
}
}
\seealso{
\code{\link[=qcc]{qcc()}}
}
\author{
Luca Scrucca
}
