% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.t2.R
\name{stats.T2}
\alias{stats.T2}
\alias{limits.T2}
\title{Statistics used in computing and drawing the Hotelling T^2 chart for subgrouped data}
\usage{
stats.T2(data, center = NULL, cov = NULL)

limits.T2(ngroups, size, nvars, conf)
}
\arguments{
\item{data}{the observed data values}

\item{center}{a vector of values to use for center of input variables.}

\item{cov}{a matrix of values to use for the covariance matrix of input
variables.}

\item{ngroups}{number of groups}

\item{size}{sample size}

\item{nvars}{number of variables}

\item{conf}{confidence level (0 < \code{conf} < 1)}
}
\value{
The function \code{stats.T2} returns a list with components:
\itemize{
\item \code{statistics}: a vector of values for the \eqn{T^2} statistic.
\item \code{means}: a matrix of within group means for each variable.
\item \code{center}: sample/group center statistic.
\item \code{S}: covariance matrix.
}

The function \code{limits.T2} returns a list with components:
\itemize{
\item \code{control}: control limits.
\item \code{prediction}: prediction limits.
}
}
\description{
These functions are used to compute statistics required by the \eqn{T^2}
chart.
}
\references{
\itemize{
\item Mason, R. L., and Young, J. C. (2002). \emph{Multivariate Statistical Process Control with Industrial Applications}. SIAM.
\item Montgomery, D.C. (2013). \emph{Introduction to Statistical Quality Control}, 7th ed. New York: John Wiley & Sons.
\item Ryan, T. P. (2011). \emph{Statistical Methods for Quality Improvement}, 3rd ed. New York: John Wiley & Sons, Inc.
}
}
\seealso{
\code{\link[=mqcc]{mqcc()}}, \code{\link[=stats.T2.single]{stats.T2.single()}}
}
\author{
Luca Scrucca
}
