% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.xbarone.R
\name{stats.xbar.one}
\alias{stats.xbar.one}
\alias{sd.xbar.one}
\alias{limits.xbar.one}
\title{Statistics used in computing and drawing a Shewhart xbar chart for one-at-time data}
\usage{
stats.xbar.one(data, sizes)

sd.xbar.one(data, sizes, std.dev = c("MR", "SD"), r = 2, ...)

limits.xbar.one(center, std.dev, sizes, nsigmas = NULL, conf = NULL)
}
\arguments{
\item{data}{The observed data values.}

\item{sizes}{samples sizes. Not needed, \code{size = 1} is used.}

\item{std.dev}{within group standard deviation. Optional for
\code{sd.xbar.one} function, required for \code{limits.xbar.one}. See details.}

\item{r}{number of successive pairs of observations for computing the
standard deviation based on moving ranges of r points.}

\item{...}{catches further ignored arguments.}

\item{center}{The sample/group center statistic.}

\item{nsigmas}{Number of sigmas used to compute control limits.
Ignored when the \code{conf} argument is provided.}

\item{conf}{Confidence level used to compute control limits.
Must be a numeric value in \eqn{(0,1)}.}
}
\value{
The function \code{stats.xbar.one} returns a list with components
\code{statistics} and \code{center}.

The function \code{sd.xbar.one} returns \code{std.dev} the standard
deviation of the statistic charted.

The function \code{limits.xbar.one} returns a matrix with lower and upper
control limits.
}
\description{
These functions are used to compute statistics required by the xbar chart
for one-at-time data.
}
\details{
Methods available for estimating the process standard deviation:
\itemize{
\item \code{"MR"}: moving range; this estimate is based on the scaled mean of moving
ranges.
\item \code{"SD"}: sample standard deviation; this estimate is defined as
\code{sd(x) / cd(n)}, where \code{n} is the number of individual measurements of
\code{x}.
}
}
\examples{
x <- antifreeze[["water"]] # See `?antifreeze`
# 1) using MR (default)
qcc(x, type="xbar.one", data.name="Water content (in ppm) of batches of antifreeze")
# 2) using SD
qcc(x, type="xbar.one", std.dev = "SD", data.name="Water content (in ppm) of batches of antifreeze")

# "as the size increases further, we would expect sigma-hat to settle down
#  at a value close to the overall sigma-hat" (Wetherill and Brown, 1991, p. 121)
sigma <- NA
k <- 2:24
for (j in k) sigma[j] <- sd.xbar.one(x, k=j)
plot(k, sigma[k], type="b")     # plot estimates of sigma for 
abline(h=sd(x), col=2, lty=2)   # different values of k
}
\references{
\itemize{
\item Montgomery, D.C. (2013). \emph{Introduction to Statistical Quality Control}, 7th ed. New York: John Wiley & Sons.
\item Ryan, T. P. (2011). \emph{Statistical Methods for Quality Improvement}, 3rd ed. New York: John Wiley & Sons, Inc.
\item Wetherill, G. B., and Brown, D. W. (1991). \emph{Statistical Process Control}. New York: Chapman & Hall.
}
}
\seealso{
\code{\link[=qcc]{qcc()}}
}
\author{
Luca Scrucca
}
