% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.xbar.R
\name{stats.xbar}
\alias{stats.xbar}
\alias{sd.xbar}
\alias{limits.xbar}
\title{Statistics used in computing and drawing a Shewhart xbar chart}
\usage{
stats.xbar(data, sizes)

sd.xbar(
  data,
  sizes,
  std.dev = c("UWAVE-R", "UWAVE-SD", "MVLUE-R", "MVLUE-SD", "RMSDF"),
  ...
)

limits.xbar(center, std.dev, sizes, nsigmas = NULL, conf = NULL)
}
\arguments{
\item{data}{The observed data values.}

\item{sizes}{samples sizes. Optional}

\item{std.dev}{within group standard deviation. Optional for \code{sd.xbar}
function, required for \code{limits.xbar}. See details.}

\item{...}{catches further ignored arguments.}

\item{center}{The sample/group center statistic.}

\item{nsigmas}{Number of sigmas used to compute control limits.
Ignored when the \code{conf} argument is provided.}

\item{conf}{Confidence level used to compute control limits.
Must be a numeric value in \eqn{(0,1)}.}
}
\value{
The function \code{stats.xbar} returns a list with components
\code{statistics} and \code{center}.

The function \code{sd.xbar} returns \code{std.dev} the standard deviation of
the statistic charted. This is based on results from Burr (1969).

The function \code{limits.xbar} returns a matrix with lower and upper
control limits.
}
\description{
These functions are used to compute statistics required by the xbar chart.
}
\details{
The following methods are available for estimating the process standard
deviation:
\itemize{
\item \code{"UWAVE-R"}: UnWeighted AVErage of within-group estimates based on
within-group Ranges.
\item \code{"UWAVE-SD"}: UnWeighted AVErage of within-group estimates based on
within-group Standard Deviations.
\item \code{"MVLUE-R"}: Minimum Variance Linear Unbiased Estimator computed as a
weighted average of within-group estimates based on within-group Ranges.
\item \code{"MVLUE-SD"}: Minimum Variance Linear Unbiased Estimator computed as a
weighted average of within-group estimates based on within-group Standard
Deviations.
\item \code{"RMSDF"}: Root-Mean-Square estimator computed as a weighted average of
within-group estimates based on within-group Standard Deviations.
}

Depending on the chart, a method may be available or not, or set as the
default according to the following table:\tabular{llll}{
   Method \tab \code{"xbar"} \tab \code{"R"} \tab \code{"S"} \cr
   \code{"UWAVE-R"} \tab default \tab default \tab not available \cr
   \code{"UWAVE-SD"} \tab available \tab not available \tab default \cr
   \code{"MVLUE-R"} \tab available \tab available \tab not available \cr
   \code{"MVLUE-SD"} \tab available \tab not available \tab available \cr
   \code{"RMSDF"} \tab available \tab not available \tab available \cr
}


Detailed definitions of formulae implemented are available in the SAS/QC
User's Guide.
}
\references{
\itemize{
\item Burr, I. W. (1969). Control charts for measurements with varying sample sizes. \emph{Journal of Quality Technology}, \strong{1}(3), 163--167.
\item Montgomery, D.C. (2013). \emph{Introduction to Statistical Quality Control}, 7th ed. New York: John Wiley & Sons.
\item Wetherill, G. B., and Brown, D. W. (1991). \emph{Statistical Process Control}. New York: Chapman & Hall.
}
}
\seealso{
\code{\link[=qcc]{qcc()}}
}
\author{
Luca Scrucca
}
