% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{viscosity}
\alias{viscosity}
\title{Viscosity for aircraft primer paint data}
\format{
A data frame with 35 observations on the following 3 variables.
\describe{
\item{batch}{batch number}
\item{viscosity}{viscosity measure}
\item{trial}{preliminary sample indicator (TRUE/FALSE)}
}
}
\source{
Montgomery, D.C. (2005) \emph{Introduction to Statistical
Quality Control}, 5th ed, New York, John Wiley & Sons, pp. 232-235
}
\usage{
viscosity
}
\description{
"The viscosity of an aircraft primer paint is an important quality
characteristic. The product is produced in batches, and because each batch
takes several hours to produce, the production rate is too slow to allow for
rational subgroups of size greater than one." (Montgomery, 2005, p. 232)
}
\seealso{
Other datasets in \code{qcc} package:
\code{\link{RyanMultivar}},
\code{\link{antifreeze}},
\code{\link{boiler}},
\code{\link{circuit}},
\code{\link{dyedcloth}},
\code{\link{orangejuice}},
\code{\link{pcmanufact}},
\code{\link{pistonrings}}
}
\concept{datasets in \code{qcc} package}
\keyword{datasets}
