% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{d3}
\alias{d3}
\title{The \eqn{d_3}{d3} Constant}
\usage{
d3(n)
}
\arguments{
\item{n}{A vector of sample sizes.}
}
\value{
A vector of calculated \eqn{d_3}{d3} constants.
}
\description{
Calculate the \eqn{d_3}{d3} constants for a vector \code{n} of sample sizes.
}
\details{
\eqn{d_3(n)}{d3(n)} is the standard deviation of the range of \code{n} observations
from a standard normal distribution.
\deqn{d_3(n) = \frac{Stdev(r)}{\sigma}}{d3(n) = stdev(r)/sigma}

Not to be confused with the \eqn{D_3}{D3}, which is
\deqn{D_4=1-3\frac{d_3}{d_2}}{D4 = 1 - 3 * (d3 / d2)}
}
\references{
\itemize{
\item Cano, Emilio L., Moguerza, Javier M. and Redchuk, Andres. 2012. \emph{Six Sigma with {R}. Statistical Engineering for ProcessImprovement}, Use R!, vol. 36. Springer, New York.
}
}
\seealso{
For other implementations in R:
\code{\link[SixSigma:ss.cc.getd3]{SixSigma::ss.cc.getd3()}}, \code{\link[IQCC:d3]{IQCC::d3()}}, \code{\link[shewhartr:shewhart_constants]{shewhartr::shewhart_constants()}}

Other constants for Shewhart charts:
\code{\link[=c4]{c4()}},
\code{\link[=d2]{d2()}},
\code{\link[=d4]{d4()}}
}
\concept{constants for Shewhart charts}
