% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqcc.R
\name{ellipseChart}
\alias{ellipseChart}
\title{Multivariate Quality Control Charts}
\usage{
ellipseChart(
  object,
  chart.all = getOption("qcc.chart.all"),
  show.id = FALSE,
  ngrid = 50,
  confidence.level,
  correct.multiple = TRUE,
  title,
  xlim,
  ylim,
  xlab,
  ylab,
  restore.par = TRUE,
  ...
)
}
\arguments{
\item{object}{an object of class \code{'mqcc'}.}

\item{chart.all}{a logical value indicating whether both statistics for
\code{data} and for \code{newdata} (if given) should be plotted.}

\item{show.id}{a logical value indicating whether to plot point labels
(\code{TRUE}) or symbols (\code{FALSE}) for group means.}

\item{ngrid}{a value for the size of the grid over which the ellipse is
evaluated.}

\item{confidence.level}{a numeric value between 0 and 1 specifying the
confidence level of the computed probability limits.}

\item{correct.multiple}{a logical value indicating whether to correct or not
for multiple comparisons.}

\item{title}{a character string specifying the main title. Set \code{title = FALSE} or \code{title = NA} to remove the title.}

\item{xlim}{a numeric vector specifying the limits for the x-axis.}

\item{ylim}{a numeric vector specifying the limits for the y-axis.}

\item{xlab}{a string giving the label for the x-axis.}

\item{ylab}{a string giving the label for the y-axis.}

\item{restore.par}{a logical value indicating whether the previous
\code{par} settings must be restored. If you need to add points, lines, etc.
to a control chart set this to \code{FALSE}.}

\item{...}{additional arguments to be passed to the generic
\code{\link[=points]{points()}} function.}
}
\description{
Plot an ellipse chart for a bivariate quality control data.
}
\examples{
# See examples in help(mqcc)
}
\references{
\itemize{
\item Mason, R. L., and Young, J. C. (2002). \emph{Multivariate Statistical Process Control with Industrial Applications}. SIAM.
\item Montgomery, D.C. (2013). \emph{Introduction to Statistical Quality Control}, 7th ed. New York: John Wiley & Sons.
\item Ryan, T. P. (2011). \emph{Statistical Methods for Quality Improvement}, 3rd ed. New York: John Wiley & Sons, Inc.
}
}
\seealso{
\code{\link[=mqcc]{mqcc()}}, \code{\link[=stats.T2]{stats.T2()}}, \code{\link[=stats.T2.single]{stats.T2.single()}}
}
\author{
Luca Scrucca
}
