% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-layer.R
\name{plot_common}
\alias{plot_common}
\title{Shared Plotting Documentation}
\arguments{
\item{xtime}{a vector of date-time values as returned by
\code{\link[=Sys.time]{Sys.time()}} and \code{\link[=Sys.Date]{Sys.Date()}}. If provided it is used
for x-axis so it must be of the same length as the statistic charted.}

\item{add.stats}{a logical value indicating whether statistics and other
information should be printed at the bottom of the chart.}

\item{chart.all}{a logical value indicating whether both statistics for
\code{data} and for \code{newdata} (if given) should be plotted.}

\item{fill}{a logical value specifying if the in-control area should be
filled with the color specified in the \code{qcc.zones} option.}

\item{title}{a character string specifying the main title. Set \code{title = NULL} to remove the title.}

\item{xlab}{a string giving the label for the x-axis.}

\item{ylab}{a string giving the label for the y-axis.}

\item{xlim}{a numeric vector specifying the limits for the x-axis.}

\item{ylim}{a numeric vector specifying the limits for the y-axis.}

\item{digits}{the number of significant digits to use.}
}
\description{
Shared Plotting Documentation
}
\keyword{internal}
