% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capability.R
\name{processCapability}
\alias{processCapability}
\alias{print.processCapability}
\alias{summary.processCapability}
\alias{plot.processCapability}
\title{Process capability analysis}
\usage{
processCapability(
  object,
  spec.limits,
  target,
  std.dev,
  nsigmas,
  confidence.level = 0.95,
  ...
)

\method{print}{processCapability}(x, digits = getOption("digits"), ...)

\method{summary}{processCapability}(object, ...)

\method{plot}{processCapability}(
  x,
  add.stats = getOption("qcc.add.stats"),
  breaks = nclass.hist,
  fill = adjustcolor(getOption("qcc.zones")$fill, alpha.f = 0.5),
  color = "white",
  title,
  xlab,
  digits = getOption("digits"),
  ...
)
}
\arguments{
\item{object}{a \code{'qcc'} object of type \code{"xbar"}}

\item{spec.limits}{a two-values vector specifying the lower and upper
specification limits. For one-sided specification limits, the value of the
missing limit must be set to \code{NA}.}

\item{target}{a value specifying the target of the process. If missing the
value from the \code{'qcc'} object is used if not \code{NULL}, otherwise the
target is set at the middle value between specification limits.}

\item{std.dev}{a value specifying the within-group standard deviation. If
not provided is taken from the \code{'qcc'} object.}

\item{nsigmas}{a numeric value specifying the number of sigmas to use. If
not provided is taken from the \code{'qcc'} object.}

\item{confidence.level}{a numeric value between 0 and 1 specifying the level
to use for computing confidence intervals.}

\item{...}{catches further ignored arguments.}

\item{x}{an object of class \code{'processCapability'}.}

\item{digits}{the number of significant digits to use.}

\item{add.stats}{a logical value indicating whether statistics and
capability indices should be added at the bottom of the chart.}

\item{breaks}{a value or a function used to select the number of bins in a
histogram. See the help for \code{\link[=nclass.scott]{nclass.scott()}} for more details.}

\item{fill, color}{values specifying the colour of the filled area and the
border used for drawing the histogram.}

\item{title}{a character string specifying the plot title. Set \code{title = NULL} to remove the title.}

\item{xlab}{a character string specifying the label for the x-axis.}
}
\value{
Invisibly returns a list with components:
\itemize{
\item \code{nobs}: number of non-missing observations.
\item \code{center}: center.
\item \code{std.dev}: standard deviation.
\item \code{target}: target.
\item \code{spec.limits}: a vector of values giving the lower specification limit
(LSL) and the upper specification limit (USL).
\item \code{indices}: a matrix of capability indices (\eqn{C_p}{C_p},
\eqn{C_{pl}}{C_pl}, \eqn{C_{pu}}{C_pu}, \eqn{C_{pk}}{C_pk},
\eqn{C_{pm}}{C_pm}) and the corresponding confidence limits.
\item \code{exp}: a vector of values giving the expected fraction, based on a normal
approximation, of the observations less than LSL and greater than USL.
\item \code{obs}: a vector of values giving the fraction of observations less than
LSL and greater than USL.
}
}
\description{
Computes process capability indices for a \code{'qcc'} object of type
\code{"xbar"} and plot the histogram.
}
\details{
This function calculates confidence limits for \eqn{C_p}{C_p} using the
method described by Chou et al. (1990). Approximate confidence limits for
\eqn{C_{pl}}{C_pl}, \eqn{C_{pu}}{C_pu}, and \eqn{C_{pk}}{C_pk} are computed
using the method in Bissell (1990). Confidence limits for
\eqn{C_{pm}}{C_pm} are based on the method of Boyles (1991); this method is
approximate and it assumes that the target is midway between the
specification limits.
}
\examples{

data(pistonrings)
diameter  = qccGroups(data = pistonrings, diameter, sample)
q  = qcc(diameter[1:25,], type="xbar", nsigmas=3)
pc  = processCapability(q, spec.limits=c(73.95,74.05))
pc
plot(pc)
plot(processCapability(q, spec.limits=c(73.95,74.05), target=74.02))
plot(processCapability(q, spec.limits=c(73.99,74.01)))
plot(processCapability(q, spec.limits = c(73.99, 74.1)))

}
\references{
\itemize{
\item Bissell, A. F. (1990). How reliable is your capability index? \emph{Applied Statistics}, \strong{39}, 331--340.
\item Boyles, R. A. (1991). The Taguchi capability index. \emph{Journal of Quality Technology}, \strong{23}, 107--126.
\item Chou, Y., Owen, D. B., and Borrego, S. A. (1990). Lower confidence limits on process capability indices. \emph{Journal of Quality Technology}, \strong{22}, 223--229.
\item Montgomery, D.C. (2013). \emph{Introduction to Statistical Quality Control}, 7th ed. New York: John Wiley & Sons.
\item Wetherill, G. B., and Brown, D. W. (1991). \emph{Statistical Process Control}. New York: Chapman & Hall.
}
}
\seealso{
\code{\link[=qcc]{qcc()}}
}
\author{
Luca Scrucca
}
