\name{qcc-deprecated}
\alias{qcc.groups}
\alias{pareto.chart}
\alias{process.capability}
\alias{oc.curves}
\alias{cause.and.effect}
\alias{qcc.overdispersion.test}

\title{Deprecated Functions in \pkg{qcc} package}

\description{
These functions are provided for compatibility with older versions of the  package \pkg{qcc}, and they will likely be removed in the future. 
}

\usage{
qcc.groups(\dots)
pareto.chart(\dots)
process.capability(\dots)
oc.curves(\dots)
cause.and.effect(\dots)
qcc.overdispersion.test(\dots)
}

\arguments{
\item{\dots}{all arguments are passed down.}
}

\seealso{\code{\link{deprecated}}}

