\name{qcc-package}
\alias{qcc-package}
\docType{package}

\title{Quality Control Charts}

\description{
Shewhart quality control charts for continuous, attribute and count data. Cusum and EWMA charts. Operating characteristic curves. Process capability analysis. Pareto chart and cause-and-effect chart. Multivariate control charts.
}

%\details{}

\note{For a nice blog post discussing the \code{qcc} package, in particular how to implement the \emph{Western Eletric Rules (WER)}, see \url{http://blog.yhathq.com/posts/quality-control-in-r.html}.}

\seealso{\code{\link{qcc}}, \code{\link{mqcc}}, \code{\link{cusum}}, \code{\link{ewma}}, \code{\link{oc.curves}}, \code{\link{process.capability}}, \code{\link{pareto.chart}}, \code{\link{cause.and.effect}}.}

\author{Luca Scrucca \email{luca@stat.unipg.it}}

\references{
Scrucca, L. (2004). qcc: an R package for quality control charting and statistical process control. \emph{R News} 4/1, 11-17.
}

\keyword{package}