\name{qcc-package}
\alias{qcc-package}
\docType{package}

\title{Quality Control Charts}

\description{
Shewhart quality control charts for continuous, attribute and count data. Cusum and EWMA charts. Operating characteristic curves. Process capability analysis. Pareto chart and cause-and-effect chart. Multivariate control charts.
}

\details{
See \href{../doc/index.html}{vignette and documentation} accompanying the package.
}

\seealso{\code{\link{qcc}}, \code{\link{mqcc}}, \code{\link{cusum}}, \code{\link{ewma}}, \code{\link{ocCurves}}, \code{\link{processCapability}}, \code{\link{paretoChart}}, \code{\link{causeEffectDiagram}}.}

\author{Luca Scrucca}

\references{
Scrucca, L. (2004). qcc: an R package for quality control charting and statistical process control. \emph{R News} 4/1, 11-17.
}

\keyword{package}
