% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcc.R
\name{QCC}
\alias{QCC}
\title{Entry Point of the 'QCC' package}
\usage{
QCC(url, jwt)
}
\arguments{
\item{url}{A string representing the base url of the QCC server instance}

\item{jwt}{A string representing user's token provided to access QCC database}
}
\description{
Function QCC(url, jwt) provides the access to the main functions of extracting data in the package
}

